+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add Feign Death ranged swing reset interaction.
+
 ## [1.1.1] - 2022-08-29
 
 ### Canged